+++ template/__init__.py
@@ -5,10 +5,82 @@
 
 """Experiment with XSPEC models.
 
+The XSPEC model library is automatically initialized when the first
+call is made, and not when the module is loaded.
+
+There are three types of symbols in this package:
+
+1. model functions, such as `apec` and `TBabs`.
+2. routines that set or get values such as the abundance
+   table (`abundance`), cross-section table (`cross_sections`),
+   cosmology (`cosmology`), and the chatter level (`chatter`).
+3. routines about the models: `info` and `list_models`.
+
+Examples
+--------
+
+What version of XSPEC is being used?
+
+>>> import xspec_models_cxc as x
+>>> x.get_version()
+'12.12.0'
+
+What models are supported (the actual list depends on the
+version of XSPEC the code was compiled against)?
+
+>>> import xspec_models_cxc as x
+>>> x.list_models()
+['SSS_ice', 'TBabs', 'TBfeo', ..., 'zwndabs', 'zxipab', 'zxipcf']
+
+Evaluate the multiplication of phabs and apec models, with their
+default parameter ranges, on the grid from 0.1 to 10 keV with a bin
+size of 1 eV. Note that the return value has one less element than the
+grid, because it is evaluated over the range egrid[0]-egrid[1],
+egrid[1]-egrid[2], ..., egrid[-2]-egrid[-1] - which for this case is
+0.100-0.101, 0.101-0.102, ..., 9.998-9.999 keV.
+
+>>> import numpy as np
+>>> from matplotlib import pyplot as plt
+>>> import xspec_models_cxc as x
+>>> egrid = np.arange(0.1, 10, 0.001)
+>>> apec = x.info('apec')
+>>> phabs = x.info('phabs')
+>>> pars_apec = [p.default for p in apec.parameters]
+>>> pars_phabs = [p.default for p in phabs.parameters]
+>>> yapec = x.apec(energies=egrid, pars=pars_apec)
+>>> yphabs = x.phabs(energies=egrid, pars=pars_phabs)
+>>> ymodel = yphabs * yapec
+>>> emid = (egrid[:-1] + egrid[1:]) / 2
+>>> plt.plot(emid, ymodel, label='phabs * apec')
+>>> plt.yscale('log')
+>>> plt.ylim(1e-9, 0.01)
+>>> plt.legend()
+>>> plt.xlabel('Energy (keV)')
+>>> plt.ylabel('Photon/cm$^2$/s')
+
+We can include a convolution component - in this case the kdblur model
+- even if it is physically unrealistic. The two differences here is
+that the model requires the data to be convolved to be sent in as the
+`model` argument, and that this array is changed by the routine (in
+the same way that the out parameter works for NumPy ufunc
+routines). The convolution models do also return the value so we could
+have said
+
+    out = x.kdblur(ebergies=.., pars=.., model=ymodel.copy())
+
+which would keep the original model values.
+
+>>> kdblur = x.info('kdblur')
+>>> pars_kdblur = [p.default for p in kdblur.parameters]
+>>> x.kdblur(energies=egrid, pars=pars_kdblur, model=ymodel)
+>>> plt.plot(emid, ymodel, alpha=0.8, label='Convolved')
+>>> plt.legend()
+
 """
 
 from dataclasses import dataclass
 from enum import Enum, auto
+from typing import List, Optional, Sequence
 
 from ._compiled import *
 
@@ -32,6 +104,32 @@ class LanguageStyle(Enum):
     F77Style8 = auto()
 
 
+class ParamType(Enum):
+    """The XSPEC parameter type."""
+
+    Default = auto()
+    Switch = auto()
+    Scale = auto()
+    Periodic = auto()
+
+
+@dataclass
+class XSPECParameter:
+    """An XSPEC parameter."""
+
+    paramtype: ParamType
+    name: str
+    default: float
+    units: Optional[str] = None
+    frozen: bool = False
+    # Would it be better to just have limits = [hardmin, softmni, softmax, hardmax]?
+    softmin: Optional[float] = None
+    softmax: Optional[float] = None
+    hardmin: Optional[float] = None
+    hardmax: Optional[float] = None
+    delta: Optional[float] = None
+
+
 @dataclass
 class XSPECModel:
     """An XSPEC model."""
@@ -42,6 +140,7 @@ class XSPECModel:
     language: LanguageStyle
     elo: float
     ehi: float
+    parameters: Sequence[XSPECParameter]
     use_errors: bool = False
     can_cache: bool = True
 
@@ -51,7 +150,7 @@ _info = {
     }
 
 
-def info(model):
+def info(model: str) -> XSPECModel:
     """Return information on the XSPEC model from the model.dat file.
 
     This returns the information on the model as taken from the XSPEC
@@ -71,6 +170,17 @@ def info(model):
     --------
     list_models
 
+    Examples
+    --------
+
+    >>> m = info('apec')
+    >>> m.name
+    'apec'
+    >>> m.modeltype
+    <ModelType.Add: 1>
+    >>> [(p.name, p.default, p.units) for p in m.parameters]
+    [('kT', 1.0, 'keV'), ('Abundanc', 1.0, None), ('Redshift', 0.0, None)]
+
     """
 
     # We want case-insensitive comparison but for the keys to retain
@@ -85,7 +195,9 @@ def info(model):
 
     return out
 
-def list_models(modeltype=None, language=None):
+# Do we need Optional here?
+def list_models(modeltype: Optional[ModelType] = None,
+                language: Optional[LanguageStyle] = None) -> List[str]:
     """Returns the names of XSPEC models from the model.dat file.
 
     This returns the information on the model as taken from the XSPEC
@@ -107,18 +219,41 @@ def list_models(modeltype=None, language=None):
     --------
     info
 
+    Notes
+    -----
+    The restrictions are combined, so setting both `modeltype` and
+    `language` will select only those models which match both filters.
+
+    Examples
+    --------
+
+    >>> len(list_models())
+    231
+
+    >>> 'tbabs' in list_models()
+    False
+    >>> 'TBabs' in list_models()
+    True
+
+    With XSPEC 12.12.0 / HEASOFT 6.29:
+
+    >>> list_models(modeltype=ModelType.Con)
+    ['cflux', 'clumin', 'cpflux', 'gsmooth', ..., 'zashift', 'zmshift']
+
+    >>> list_models(modeltype=ModelType.Con, language=LanguageStyle.F77Style4)
+    ['kyconv', 'rgsxsrc', 'thcomp']
+
     """
 
     out = set()
     for k, v in _info.items():
-        if modeltype is None and language is None:
-            out.add(k)
+
+        if modeltype is not None and v.modeltype != modeltype:
             continue
 
-        if modeltype is not None and v.modeltype == modeltype:
-            out.add(k)
+        if language is not None and v.language != language:
+            continue
 
-        if language is not None and v.language == language:
-            out.add(k)
+        out.add(k)
 
     return sorted(out)